% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalAnalysis.R
\name{causalAnalysis}
\alias{causalAnalysis}
\title{causalAnalysis}
\usage{
causalAnalysis(
  df,
  bioCols,
  ind_cov,
  metaCols,
  patientCol,
  depVar = NULL,
  colorVar = NULL,
  REML = FALSE,
  parametric = TRUE,
  exportPDF = TRUE,
  printMedTable = TRUE,
  resultsFolder = "results"
)
}
\arguments{
\item{df}{A given dataframe}

\item{bioCols}{Range of columns with biomarker information (e.g., 16:300)}

\item{ind_cov}{Individual covariates for which we want to test the interaction
with the biomarkers. They will be included as fixed effects in all models.}

\item{metaCols}{Range of columns with metadata information (e.g., 2:15)}

\item{patientCol}{Name of the ID or patient column in the top-table.}

\item{depVar}{Dependent variable, as clinical outcome or score (NULL by default)}

\item{colorVar}{A categorical variable from the dataframe to color the points.
(e.g., "Group"). NULL by default.}

\item{REML}{Algorithm used for estimating the model parameters. It is recommended
to set it as FALSE (it will use ML) if we are comparing models.}

\item{parametric}{Option to perform parametric tests (Student's t, ANOVA)
or non-parametric (Wilcoxon or Kruskal-Wallis) in boxplots. TRUE by default.}

\item{exportPDF}{Export plots to a PDF (TRUE by default; if FALSE, plots will
be printed).}

\item{printMedTable}{Print the table with the information about the biomarkers
that lost significance after adding a covariate (mediation analysis) (TRUE
by default).}

\item{resultsFolder}{Name of the directory in which the results will be stored
("results" by default).}
}
\value{
A dataframe with the results of the different linear models.
}
\description{
This function runs all the functions in the package.
}
