% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covBarplot.R
\name{covBarplot}
\alias{covBarplot}
\title{covBarplot}
\usage{
covBarplot(df, metaCols, exportPDF = TRUE, resultsFolder = "results")
}
\arguments{
\item{df}{A provided dataframe.}

\item{metaCols}{Range of columns with metadata information (e.g., 2:15).}

\item{exportPDF}{Export plots to a PDF (TRUE by default; if FALSE, plots will
be printed).}

\item{resultsFolder}{Name of the directory in which the results will be stored
("results" by default).}
}
\value{
A table with a summary of the distribution of the categorical
variables of the dataframe depending on the other categorical variables.
}
\description{
This function generates several bar plots that analyze the distribution of
categorical variables of a given dataframe with respect to the other
categorical variables. In addition, it performs a Chi-squared test to
compare if the differences between the two categorical variables tested
in each plot are significant or not. It provides the p-value of the test
as the subtitle, as well as the sample size for each group. It takes only
three arguments: the dataframe that we want to analyze ("df"), the range of
columns in which we have the covariates information ("metaCols") and an
optional argument called "exportPDF" (NULL by default) that takes the
name of the PDF file that will store the plots. Finally, the function returns
a table with a summary of the distribution of categorical variables depending
on the other categorical variables.
}
