% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covBoxplot.R
\name{covBoxplot}
\alias{covBoxplot}
\title{covBoxplot}
\usage{
covBoxplot(
  df,
  metaCols,
  parametric = TRUE,
  exportPDF = TRUE,
  resultsFolder = "results"
)
}
\arguments{
\item{df}{A given dataframe.}

\item{metaCols}{Range of columns with metadata information (e.g., 2:15).}

\item{parametric}{Option to perform parametric tests (Student's t, ANOVA)
or non-parametric (Wilcoxon or Kruskal-Wallis). TRUE by default.}

\item{exportPDF}{Export plots to a PDF (TRUE by default; if FALSE, plots will
be printed).}

\item{resultsFolder}{Name of the directory in which the results will be stored
("results" by default).}
}
\value{
A table with a summary of the distribution of numeric variables
depending on categorical variables
}
\description{
This function generates several boxplots that analyze the distribution of
numerical variables of a given dataframe with respect to the categorical
variables. It also performs parametric (Student's t or ANOVA depending
on the levels of the categorical variable) or non-parametric tests (Wilcoxon
or Kruskal-Wallis) to assess if the differences between the distributions
are significant. It shows in the boxplot's subtitle the name of the test
that has been used, the p-value and the sample size for each level of the
categorical variable. The function takes four arguments: the dataframe ("df"),
the range of columns with metadata information ("metaCols"), the option to
perform parametric or non-parametric tests ("parametric") and an
optional argument called "exportPDF" (NULL by default) that takes the
name of the PDF file that will store the plots. Finally, the function returns
a table with a summary of the distribution of numeric variables depending
on the levels of the categorical variables.
}
