% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covPairs.R
\name{covPairs}
\alias{covPairs}
\title{covPairs}
\usage{
covPairs(df, metaCols, colorVar = NULL)
}
\arguments{
\item{df}{A provided dataframe.}

\item{metaCols}{Range of columns in which we have the metadata (e.g., 2:20)}

\item{colorVar}{A categorical variable from the dataframe to color the points.
(e.g., "Group"). NULL by default.}
}
\value{
A pairs plot.
}
\description{
This function is an adaptation of the ggpairs function from the GGally package
that eases its use with any given dataframe. It needs a dataframe, the range
of the metadata columns and an optional grouping variable to color the points
and plots. From the range of the metadata columns, the function extracts
the numeric variables and generates the plot.
}
