% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeBioCov.R
\name{lmeBioCov}
\alias{lmeBioCov}
\title{lmeBioCov}
\usage{
lmeBioCov(
  df,
  bio_names,
  ind_cov,
  covariate,
  patientCol,
  fixedFormula,
  randomFormula,
  REML
)
}
\arguments{
\item{df}{A given dataframe}

\item{bio_names}{Vector with the names of the biomarkers to test.}

\item{ind_cov}{Individual covariates for which we want to test the interaction
with the biomarkers. They will be included as fixed effects in all models.}

\item{covariate}{The additional covariate that we are including in the model.}

\item{patientCol}{Name of the ID or patient column in the top-table.}

\item{fixedFormula}{Formula of the Fixed Effects of the Model.}

\item{randomFormula}{Formula of the Random Effects of the Model.}

\item{REML}{Algorithm used for estimating the model parameters. It is recommended
to use FALSE (it will use ML) if we are comparing models.}
}
\value{
A dataframe with the results of the different linear models.
}
\description{
This function fits linear mixed-effects models to test the influence of a
set of covariates in the levels of the dependent variable (biomarkers). It takes
as input a dataframe ("df") containing the data to be analyzed, a vector with
the names of the biomarkers ("bio_names") to be tested, a vector of individual
covariates ("ind_cov") in which we will test the interaction, the covariate
("cov") that we are adding to the model besides the individual covariates,
the name of the ID or patient column in the top-table ("patientCol"), the
formula of the fixed effects of the model ("fixedFormula"), the formula of the
random effects of the model ("randomFormula"), and the algorithm used for
estimating the model parameters (REML). The function returns a dataframe
with the results of the different linear models for each biomarker, including
the coefficients, standard errors, t-values, p-values, R-squared, AIC, and
the number of outliers.
}
