% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeBioLRT.R
\name{lmeBioLRT}
\alias{lmeBioLRT}
\title{lmeBioLRT}
\usage{
lmeBioLRT(
  df,
  bio_names,
  covariates,
  patientCol,
  ind_cov,
  fixedFormula1,
  fixedFormula2,
  randomFormula,
  REML
)
}
\arguments{
\item{df}{A given dataframe}

\item{bio_names}{Vector with the names of the biomarkers to test.}

\item{covariates}{Vector with the covariates that we want to include in the model.}

\item{patientCol}{Name of the ID or patient column in the dataframe.}

\item{ind_cov}{Individual covariates for which we want to test the interaction
with the biomarkers. They will be included as fixed effects in all models.}

\item{fixedFormula1}{Formula of the first Fixed Effects of the Model
(with the covariate added to the interaction).}

\item{fixedFormula2}{Formula of the second Fixed Effects of the Model
(without the covariate added to the interaction).}

\item{randomFormula}{Formula of the Random Effects of both Models.}

\item{REML}{Algorithm used for estimating the model parameters. It is recommended
to set it as FALSE (it will use ML) if we are comparing models.}
}
\value{
A dataframe with the summary of the model comparison.
}
\description{
This function performs a likelihood ratio test (LRT) to compare the fit of two
linear mixed-effect models for multiple biomarkers in a given dataframe. The
function uses the biomarkers as the dependent variable and it compares
the differences between a model that adds a covariate to a given interaction
and the model with the covariate added to the model but outside of the
main interaction. The function takes in a dataframe "df" and several other
parameters, including "bio_names" (a vector with the names of the biomarkers to
test), "covariates" (a vector with the covariates to include in the model),
"patientCol" (the name of the ID or patient column in the df), "ind_cov"
(the individual covariates for which the interaction with the biomarkers will
be tested and included as fixed effects in all models), "fixedFormula1" and
"fixedFormula2" (the formulas for the first and second fixed effects of the
model), "randomFormula" (the formula for the random effects of both
models), and "REML" (the algorithm used for estimating the model parameters).
The function first retrieves the indexes of the biomarkers, patient, and
individual covariates and creates a data frame to store the data. It loops
through all the biomarkers, generates two models for each one, performs the
LRT, and stores the relevant data in a data frame. The function exports the
results to an Excel file and returns the data frame with the comparisons.
The function uses the lme function from the nlme package to fit the
mixed-effects models.
}
