% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationAnalysis.R
\name{mediationAnalysis}
\alias{mediationAnalysis}
\title{mediationAnalysis}
\usage{
mediationAnalysis(
  df,
  metaCols,
  patientCol,
  bio_names,
  ind_cov,
  depVar,
  REML,
  printFormula = TRUE,
  printMedTable = TRUE,
  resultsFolder = "results"
)
}
\arguments{
\item{df}{A given dataframe}

\item{metaCols}{Range of columns with metadata information (e.g., 2:15)}

\item{patientCol}{Name of the ID or patient column in the top-table.}

\item{bio_names}{Vector with the names of the biomarkers to test.}

\item{ind_cov}{Individual covariates for which we want to test the interaction
with the biomarkers. They will be included as fixed effects in all models.}

\item{depVar}{Dependent variable, as clinical outcome or score.}

\item{REML}{Algorithm used for estimating the model parameters. It is recommended
to set it as FALSE (it will use ML) if we are comparing models.}

\item{printFormula}{Print the formula of the different models in the console
(TRUE by default).}

\item{printMedTable}{Print the table with the information about the biomarkers
that lost significance after adding a covariate (mediation analysis) (TRUE
by default).}

\item{resultsFolder}{Name of the directory in which the results will be stored
("results" by default).}
}
\value{
A dataframe with the results of the mediation analysis
}
\description{
This function mediationAnalysis performs a mediation analysis between the biomarkers
and the dependent variable while considering the effect of individual covariates
on this relationship. It takes as input a data frame ("df"), the names of the
columns corresponding to metadata ("metaCols"), the column name containing the
patients' IDs ("patientCol"), the names of the columns corresponding to
biomarkers ("bioCols"), the individual covariates ("ind_cov"), the dependent
variable ("depVar"), the method used to perform the analysis ("REML"),
and a logical argument to indicate whether the formula of the model should be
printed ("printFormula", TRUE by default).
}
\details{
First, the function identifies the biomarkers with significant interactions with
the individual covariates, and then, it performs again the same analysis but
adding the other covariates that the user wants to test. The results are stored
in a data frame and written to an Excel file in the "results" folder. The
function returns a dataframe with the biomarkers which interaction with the individual
covariates is significant in a model with the dependent variable.
}
