% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalBoxplot.R
\name{pvalBoxplot}
\alias{pvalBoxplot}
\title{pvalBoxplot}
\usage{
pvalBoxplot(
  df,
  covariates,
  pvalCol,
  covCol,
  exportPDF = TRUE,
  sigs,
  resultsFolder = "results"
)
}
\arguments{
\item{df}{A dataframe with the results of a linear model.}

\item{covariates}{Vector with the covariates that we want to include in the model.}

\item{pvalCol}{Name of the p-value column in the dataframe.}

\item{covCol}{Name of the covariates column in the dataframe.}

\item{exportPDF}{Name of the PDF file in which the boxplot is exported to
(NULL by default; if NULL, it will print the plot).}

\item{sigs}{Vector with the desired adjusted p-value significance levels
(e.g., c(0.01, 0.05)).}

\item{resultsFolder}{Name of the directory in which the results will be stored
("results" by default).}
}
\value{
A table with the number of biomarkers below the desired significance
thresholds (adjusted p-value) for each covariate.
}
\description{
This function generates a boxplot and a table summarizing the number of
significant biomarkers by covariate after running a linear model. Besides the
"df", it takes as input a vector of "covariates" (the vector with the covariates
that we will put into the model) "pvalCol" (indicating the name of the p-value
column), and a "covCol" (which indicates the name of the Column containing the
covariate information). The function also has an optional argument, "exportPDF" (which specifies the file name to which
the plot should be exported as a PDF file), and "sigs" (which is a vector of
significance levels to be used in the output table).
}
